评教系统最终版/评教系统教师端/backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
redis
python-jose[cryptography]
python-multipart
python-dotenv
pydantic-settings
websockets
pandas
scikit-learn
python-docx
reportlab
python-magic
